+++ docs/components/Disclosure/Mobile.stories.tsx
@@ -1,12 +1,15 @@
 import React, { useState } from "react";
 import { ComponentMeta, ComponentStory } from "@storybook/react";
-import { Disclosure, Text } from "@jobber/components-native";
+import { Disclosure } from "@jobber/components-native";
 
 export default {
   title: "Components/Layouts and Structure/Disclosure/Mobile",
   component: Disclosure,
   parameters: {
     viewMode: "story",
+    previewTabs: { code: { hidden: false } },
+    viewport: { defaultViewport: "mobile1" },
+    showNativeOnWebDisclaimer: true,
   },
 } as ComponentMeta<typeof Disclosure>;
 
@@ -17,7 +20,7 @@ const BasicTemplate: ComponentStory<typeof Disclosure> = args => {
 
 export const Basic = BasicTemplate.bind({});
 Basic.args = {
-  header: <Text>Advanced Instructions</Text>,
+  header: "Advanced Instructions",
   content: "For every 2 team members you add, your profits will triple.",
   isEmpty: false,
 };